Refactor: Updated controller and impl class to return Response

+++ src/main/java/com/bridgelabz/emppayrollapp/service/implementation/EmployeeImpl.java
@@ -1,12 +1,15 @@
 package com.bridgelabz.emppayrollapp.service.implementation;
 
+import java.util.ArrayList;
 import java.util.List;
 
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import com.bridgelabz.emppayrollapp.dto.EmployeeDto;
+import com.bridgelabz.emppayrollapp.exception.EmployeeException;
 import com.bridgelabz.emppayrollapp.model.Employee;
 import com.bridgelabz.emppayrollapp.repo.EmpRepository;
 import com.bridgelabz.emppayrollapp.service.IEmployee;
@@ -18,46 +21,68 @@ public class EmployeeImpl implements IEmployee {
 	@Autowired
 	private EmpRepository empRepository;
 	
-//	@Autowired
-//	private ModelMapper mappy;
+	@Autowired
+	private ModelMapper modelMap;
 	
-//	@Override
-//	public Response addEmployee(EmployeeDto empDto) {
-//		Employee employee= mappy.map(empDto, Employee.class);
-//		empRepository.save(employee);
-//		System.out.println("Added to repo.");
-//		return new Response(200, "Added sucessfully.");
-//	}
-
 	@Override
-	public Response addEmployee(Employee emp) {
+	public Response addEmployee(EmployeeDto empDto) {
+		
+		Employee employeeExist = empRepository.findByName(empDto.getName());
+		if(employeeExist!=null) {
+			throw new EmployeeException(HttpStatus.BAD_REQUEST.value(), "Record Already Exists!");
+		}
+		
+		Employee emp = modelMap.map(empDto, Employee.class);
 		empRepository.save(emp);
-		System.out.println("Added to repository.");
+		System.out.println("Added to repo.");
 		return new Response(200, "Added sucessfully.");
 	}
 
 	@Override
-	public List<Employee> getEmployee() {
-		System.out.println(empRepository.findAll());
-		return empRepository.findAll();
+	public Response getEmployee() {
+		List<Employee> empRecord=empRepository.findAll();
+		if(empRecord.isEmpty()) {
+			throw new EmployeeException(HttpStatus.NO_CONTENT.value(), "No record Exists");
+		}
+		System.out.println(empRecord);
+		List<EmployeeDto> dtoRecord=new ArrayList<>();
+		for (Employee emp : empRecord) {
+			EmployeeDto empDto=modelMap.map(emp, EmployeeDto.class);
+			dtoRecord.add(empDto);
+		}
+		return new Response(HttpStatus.OK.value(), dtoRecord.toString());
 	}
 
 	@Override
-	public Employee getEmployeeByID(Long id) {
-		return empRepository.findById(id).get();
+	public Response getEmployeeByID(Long id) {
+		Employee emp=empRepository.findById(id).get();
+		if (emp==null) throw new EmployeeException(HttpStatus.NO_CONTENT.value(), "Employee Doesnt Exist");
+		else return new Response(HttpStatus.OK.value(), emp.toString());
 	}
 
 	@Override
-	public void deleteEmployee(Long id) {
-		System.out.println("Del from repository.");
-		empRepository.deleteById(id);
+	public Response deleteEmployee(Long id) {
+		Employee emp=empRepository.findById(id).get();
+		if (emp==null) throw new EmployeeException(HttpStatus.NO_CONTENT.value(), "Employee Doesnt Exist");
+		else {
+			empRepository.deleteById(id);
+			System.out.println("Del from repository.");
+		}
+		return new Response(HttpStatus.OK.value(), "Daleted successfully!");
 	}
 
 	@Override
-	public void update(Employee newEmp,Long id) {
-		this.deleteEmployee(id);
-		newEmp.setId(id);
-		empRepository.save(newEmp);
-		System.out.println("Updated to repository.");
+	public Response update(EmployeeDto newEmpdto,Long id) {
+		
+		Employee emp=empRepository.findById(id).get();
+		if (emp==null) throw new EmployeeException(HttpStatus.NO_CONTENT.value(), "Employee Doesnt Exist");
+		else {
+			Employee newEmp=modelMap.map(newEmpdto, Employee.class);
+			this.deleteEmployee(id);
+			newEmp.setId(id);
+			empRepository.save(newEmp);
+			System.out.println("Updated to repository.");
+		}
+		return new Response(HttpStatus.OK.value(), "Updated successfully!");
 	}
 }
